serverless/requirements.txt
amazon-textract-caller==0.0.24
boto3==1.24.14
botocore==1.27.14
requests==2.28.0
numpy
amazon-textract-response-parser
tabulate
pdf2image